+++ next/src/routes/_layout/questionnaire.$questionnaireId/_layout-q/releases/new.tsx
@@ -1,9 +1,16 @@
+import { useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
 
 import { questionnaireDetailsQueryOptions } from '@/api/questionnaireDetails'
+import {
+  pendingReleasesQueryOptions,
+  releasesQueryOptions,
+} from '@/api/releases'
+import { latestVersionQueryOptions } from '@/api/versions'
 import ErrorComponent from '@/components/layout/ErrorComponent'
 import CreateRelease from '@/components/release/create/CreateRelease'
 import CreateReleaseLayout from '@/components/release/create/CreateReleaseLayout'
+import { hasReleaseForVersion } from '@/components/release/overview/utils/utils'
 
 export const Route = createFileRoute(
   '/_layout/questionnaire/$questionnaireId/_layout-q/releases/new',
@@ -18,6 +25,11 @@ export const Route = createFileRoute(
     const questionnaireDetails = await queryClient.fetchQuery(
       questionnaireDetailsQueryOptions(questionnaireId),
     )
+    await Promise.all([
+      queryClient.ensureQueryData(releasesQueryOptions(questionnaireId)),
+      queryClient.ensureQueryData(pendingReleasesQueryOptions(questionnaireId)),
+      queryClient.ensureQueryData(latestVersionQueryOptions(questionnaireId)),
+    ])
     return {
       questionnaireDetails,
     }
@@ -28,6 +40,22 @@ function RouteComponent() {
   const questionnaireId = Route.useParams().questionnaireId
 
   const { questionnaireDetails } = Route.useLoaderData()
+  const { data: releases = [] } = useSuspenseQuery(
+    releasesQueryOptions(questionnaireId),
+  )
+  const { data: pendingRequests = [] } = useSuspenseQuery(
+    pendingReleasesQueryOptions(questionnaireId),
+  )
+  const { data: latestVersion } = useSuspenseQuery(
+    latestVersionQueryOptions(questionnaireId),
+  )
+
+  const latestVersionId = latestVersion?.id
+  const isPublishDisabled = hasReleaseForVersion(
+    latestVersionId,
+    releases,
+    pendingRequests,
+  )
 
   return (
     <CreateReleaseLayout>
@@ -35,6 +63,7 @@ function RouteComponent() {
         questionnaireId={questionnaireId}
         targetModes={questionnaireDetails.targetMode}
         serie={questionnaireDetails.dataCollection?.serie}
+        isPublishDisabled={isPublishDisabled}
       />
     </CreateReleaseLayout>
   )